smp_27feb/requirements.txt
django==2
django-filter
django-crispy-forms
django-widget-tweaks